+++ tests/integration/test_projectconfigs.py
@@ -405,6 +405,9 @@ def test_get_global_config(mini_sentry, relay):
     packed, signature = SecretKey.parse(relay.secret_key).pack(body)
     data = get_response(relay, packed, signature, version="3")
 
+    global_extraction_config = data["global"].pop("metricExtraction")
+    assert "span_metrics" in global_extraction_config["groups"]
+
     assert data["global"] == mini_sentry.global_config