+++ py/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Expose cardinality limit config normalization. ([3489](https://github.com/getsentry/relay/pull/3489))
+- Consider "Bearer" (case-insensitive) a password. PII will scrub all strings matching that substring. ([#3484](https://github.com/getsentry/relay/pull/3484))
+
 ## 0.8.58
 
 - Add ability to change the json parser. ([#3480](https://github.com/getsentry/relay/pull/3480))